Bump elixir to 1.1.x and drop rebar support
v. 0.0.2

To make things simpler rebar support is dropped and only rebar3 is
supported.

Before we asked only for elixir version to be ~> 1.0, which meant that
at some point a new version of Logger was introduced using the function
`bare_log` for logging (think this was around 1.1.0). That meant the
plugins compiled with 1.1.x where no longer compatible with the 1.0.4
elixir version loaded by the `vmq_elixir` plugin.

Therefore we now explicitly use the same version of Elixir in
`vmq_elixir` and `vmq_elixir_plugin`.